requirements.txt
streamlit
ultralytics
opencv-python-headless
matplotlib
Pillow
